refactor: split entity event compilation from the accepted event contract
entity-event.ts held both the writer side (compiling a command into an
event bundle) and the contract every reader replays (event shapes,
validation, write plans, owned-content manifests), and crossed the
1100-line source limit.

Move the six compilers, their bundle types and the envelope/declaration
helpers only they use into entity-event-compile.ts. The new module
imports the contract, never the reverse, so the offline-maintenance
closure rooted at ledger-backup.ts keeps its pinned membership and no
longer loads the compilers at all.
