+++ packages/kernel/test/contracts/execution-review-entity.test.ts
@@ -1,7 +1,7 @@
 // harness-test-tier: fast
 import assert from "node:assert/strict";
 import test from "node:test";
-import { compileEntityUpsert } from "../../src/domain/entity-event.ts";
+import { compileEntityUpsert } from "../../src/domain/entity-event-compile.ts";
 import {
   interpretEmbeddedEntityProjections,
   type EntityProjectionContract,
