+++ packages/kernel/test/store/legacy-entity-declaration-follower.integration.test.ts
@@ -4,7 +4,8 @@ import { mkdtempSync, readFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import path from "node:path";
 import test from "node:test";
-import { compileEntityUpsert, ownedContentForDeclarationEvent } from "../../src/domain/entity-event.ts";
+import { compileEntityUpsert } from "../../src/domain/entity-event-compile.ts";
+import { ownedContentForDeclarationEvent } from "../../src/domain/entity-event.ts";
 import type { CanonicalEventV1 } from "../../src/domain/doc-sync.contract.ts";
 import { openSqliteEventStore, sqliteLedgerPath } from "../../src/store/sqlite-event-store.ts";
 import { makeTaskEventStore, readCertifiedGitFollower } from "../../src/store/task-event-store-factory.ts";
